pyproject.toml
[project]
name = "load_balancer"
version = "0.1.0"
description = ""
authors = [
    { name = "Your Name", email = "you@example.com" },
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "asyncpg (>=0.30.0,<0.31.0)",
    "fastapi (==0.115.0)",
    "yarl (>=1.20.1,<2.0.0)",
    "uvicorn (>=0.35.0,<0.36.0)",
    "pydantic-settings (>=2.10.1,<3.0.0)",
    "tortoise-orm (>=0.25.1,<0.26.0)",
    "aerich[toml] (>=0.9.1,<0.10.0)",
    "pytest[asyncio] (>=8.4.1,<9.0.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "asgi-lifespan (>=2.1.0,<3.0.0)",
    "pytest-asyncio (>=1.1.0,<2.0.0)",
    "black (>=25.1.0,<26.0.0)",
    "anyio (>=4.10.0,<5.0.0)",
]

[tool.aerich]
tortoise_orm = "db.config.TORTOISE_CONFIG"
location = "./db/migrations"
src_folder = "./."

[build-system]
requires = [
    "poetry-core>=2.0.0,<3.0.0",
]
build-backend = "poetry.core.masonry.api"
